+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/response/bucket/TestOMBucketDeleteResponse.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.apache.hadoop.ozone.om.response;
+package org.apache.hadoop.ozone.om.response.bucket;
 
 import java.util.UUID;
 
@@ -32,23 +32,27 @@ import org.apache.hadoop.ozone.om.OMConfigKeys;
 import org.apache.hadoop.ozone.om.OMMetadataManager;
 import org.apache.hadoop.ozone.om.OmMetadataManagerImpl;
 import org.apache.hadoop.ozone.om.helpers.OmBucketInfo;
+import org.apache.hadoop.ozone.om.response.TestOMResponseUtils;
 import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos;
 import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos
     .CreateBucketResponse;
+import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos
+    .DeleteBucketResponse;
 import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos
     .OMResponse;
 import org.apache.hadoop.utils.db.BatchOperation;
 
 /**
- * This class tests OMBucketCreateResponse.
+ * This class tests OMBucketDeleteResponse.
  */
-public class TestOMBucketCreateResponse {
+public class TestOMBucketDeleteResponse {
 
   @Rule
   public TemporaryFolder folder = new TemporaryFolder();
 
   private OMMetadataManager omMetadataManager;
   private BatchOperation batchOperation;
+
 
   @Before
   public void setup() throws Exception {
@@ -72,15 +76,22 @@ public class TestOMBucketCreateResponse {
             .setCreateBucketResponse(
                 CreateBucketResponse.newBuilder().build()).build());
 
+    OMBucketDeleteResponse omBucketDeleteResponse =
+        new OMBucketDeleteResponse(volumeName, bucketName,
+            OMResponse.newBuilder()
+                .setCmdType(OzoneManagerProtocolProtos.Type.DeleteBucket)
+                .setStatus(OzoneManagerProtocolProtos.Status.OK)
+                .setDeleteBucketResponse(
+                    DeleteBucketResponse.getDefaultInstance()).build());
+
     omBucketCreateResponse.addToDBBatch(omMetadataManager, batchOperation);
+    omBucketDeleteResponse.addToDBBatch(omMetadataManager, batchOperation);
 
     // Do manual commit and see whether addToBatch is successful or not.
     omMetadataManager.getStore().commitBatchOperation(batchOperation);
 
-    Assert.assertEquals(omBucketInfo,
-        omMetadataManager.getBucketTable().get(
+    Assert.assertNull(omMetadataManager.getBucketTable().get(
             omMetadataManager.getBucketKey(volumeName, bucketName)));
   }
-
 
 }
